[update]usersこんとろーらを作成。そのた調整。
modelsにstructたちを統一
controllerの命名を規則
