+++ controller/user_controller.go
@@ -1,0 +1,8 @@
+package controller
+
+import (
+	"github.com/gin-gonic/gin"
+)
+
+func RegisterUser(c *gin.Context) {
+}
