--- model/dbope.go
@@ -1,3 +1,0 @@
-package model
-
-// 複数テーブルにわたる処理やテーブル自体への処理
