+++ model/modelpe.go
@@ -1,0 +1,3 @@
+package model
+
+// テーブルに対する処理
