+++ model/models.go
@@ -14,4 +14,15 @@ type User struct { // typeで型の定義, structは構造体
 	GroupNo     int    `gorm:"type:int(10);not null"`       // 部署番号
 }
 
-// 処理
+// 会社テーブル
+type Company struct { // typeで型の定義, structは構造体
+	CompanyNo   int    `gorm:"primary_key;AUTO_INCREMENT"` // 会社番号
+	CompanyName string `gorm:"size:20;not null"`           // 会社名
+}
+
+// 部署テーブル
+type Kgroup struct { // typeで型の定義, structは構造体
+	KgroupNo   int    `gorm:"primary_key;AUTO_INCREMENT"` // 部署番号
+	KgroupName string `gorm:"size:20;not null"`           // 部署名
+	CompanyNo  int    `gorm:"not null"`                   // 会社番号
+}
